configs: add support for PandaBoard and PandaBoard ES
This is a default configuration for Panda and PandaES, tested on both
platforms.

DEVTMPFS is enabled, to use static dev configuration one would need
to update the generic dev table for ttyO driver.

Panda is well supported in mainline kernel with omap2plus_defconfig,
so this should be safe.

U-boot SPL support is enabled by default as x-loader is now
deprecated.

With OMAP platforms u-boot,img format is prefered now, so it's enabled
by default as well.

Signed-off-by: Nicolas Dechesne <n-dechesne@ti.com>